test: enable LanguageManager-test unit and integ
