+++ test/UnitTestSuite.js
@@ -52,6 +52,8 @@ define(function (require, exports, module) {
     require("spec/JSUtils-test");
     require("spec/JSUtils-integ-test");
     require("spec/KeyBindingManager-test");
+    require("spec/LanguageManager-test");
+    require("spec/LanguageManager-integ-test");
     // require("spec/CodeHint-test");
     // require("spec/CodeInspection-test");
     // require("spec/CSSInlineEdit-test");
@@ -64,7 +66,6 @@ define(function (require, exports, module) {
     // require("spec/FindInFiles-test");
     // require("spec/InlineEditorProviders-test");
     // require("spec/InstallExtensionDialog-test");
-    // require("spec/LanguageManager-test");
     // require("spec/LanguageTools-test");
     // require("spec/LiveDevelopment-test");
     // require("spec/LiveDevelopmentMultiBrowser-test");
